+++ rs-matter/tests/im/mod.rs
@@ -19,4 +19,5 @@ mod client_invokes;
 mod client_reads;
 mod client_subscribes;
 mod client_writes;
+#[cfg(feature = "persistent-subscriptions")]
 mod subscription_reboot;